+++ source/Databricks/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # Databricks Release Notes
 
+## Version 3.0.0
+
+- Replaced `ExecuteAsync(string)` with `ExecuteAsync(string, List<SqlStatementParameter>)` for the possibility to use Parameter Markers in SQL statements.
+  - Optionally, it is possible to call the method with an empty list as parameters and it would work as before. (Still recommended to use Parameter Markers in queries to avoid SQL injection)
+
 ## Version 2.0.0
 
 - See [Version 2.0.0 release notes](./version_2_0_0.md)